patient-service/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.5.7
click==8.1.6
dnspython==2.4.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
fastapi-pagination==0.12.6
fhir.resources==7.0.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
motor==3.2.0
names==0.3.0
pydantic==1.10.11
pydantic_core==2.3.0
pymongo==4.4.1
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
